requirements.txt
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
audioread==3.0.1
boto3==1.34.2
botocore==1.34.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
docker==7.0.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.95.2
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
google-pasta==0.2.0
h11==0.14.0
h5py==3.10.0
idna==3.6
importlib-metadata==6.11.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mido==1.3.0
mpmath==1.3.0
msgpack==1.0.7
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathos==0.3.1
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
s3transfer==0.9.0
sagemaker==2.200.1
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tblib==1.7.0
threadpoolctl==3.2.0
torch==2.1.1
torchaudio==2.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.22.0
wandb==0.16.1
wcwidth==0.2.12
zipp==3.17.0
